Add custom css to separator
First thanks for the nice component. 
I was evaluating the component to use it in our site and needed to customize the separator for the css, noticed that you are already have a way to do that for some of the components (like panel and layout) so did I the same for the separator. thought to share if you are interested.
Thanks
Mohamed